fix more lint issues
