+++ clients/admin-ui/src/features/system/SystemInformationForm.tsx
@@ -13,7 +13,7 @@ import { Form, Formik, FormikHelpers } from "formik";
 import { useMemo } from "react";
 import * as Yup from "yup";
 
-import { useAppDispatch,useAppSelector } from "~/app/hooks";
+import { useAppDispatch, useAppSelector } from "~/app/hooks";
 import {
   CustomFieldsList,
   useCustomFields,
